+++ src/util/misc.rs
@@ -35,7 +35,7 @@ mod message_signing {
     use secp256k1;
     use secp256k1::recovery::{RecoveryId, RecoverableSignature};
 
-    use util::key::PublicKey;
+    use util::ecdsa::PublicKey;
     use util::address::{Address, AddressType};
 
     /// An error used for dealing with Bitcoin Signed Messages.
