+++ src/util/mod.rs
@@ -16,7 +16,7 @@
 //!
 //! Functions needed by all parts of the Bitcoin library
 
-pub mod key;
+pub mod ecdsa;
 pub mod address;
 pub mod amount;
 pub mod base58;
@@ -30,6 +30,7 @@ pub mod psbt;
 pub mod taproot;
 pub mod uint;
 pub mod bip158;
+pub mod key;
 
 pub(crate) mod endian;
